test(sig-events): split api test ownership
